Move function + downgrade `antd` by one version
* `antd` was downgraded by one version because it appears there is a small bug in the Table filter dropdown + sorter functionality.
